+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added parallel processing for FedEx API
+
 ## [1.17.1] - 2022-06-14
 
 ### Added